Centralize atomic JSON persistence
